refactor(handle): give JobHandle the handle module; move the task wrapper to task.rs
Resolves the "handle" overload: the public marquee type `JobHandle` now lives
in `handle.rs` (was the awkward `handle_api.rs`), and the internal RAII tokio
JoinHandle wrapper `OwnedTaskHandle` moves to `task.rs` (was `handle.rs`). Each
file is now named for its primary type; no type renames, no public API change.

Also fixes stale rustdoc links surfaced by `cargo doc` (pre-existing, non-fatal):
the removed `Jobs::await_completion` now points at `JobHandle::await_completion`,
and the public `JobSnapshot`/`Job` docs no longer intra-doc-link the crate-private
`JobExecutionRow`.
