+++ src/lib.rs
@@ -210,7 +210,6 @@ mod current;
 mod dispatcher;
 mod entity;
 mod handle;
-mod handle_api;
 mod job_execution;
 mod migrate;
 mod notification_router;
@@ -222,6 +221,7 @@ mod repo;
 mod runner;
 mod snapshot;
 mod spawner;
+mod task;
 mod tracker;
 
 pub mod error;
@@ -235,7 +235,7 @@ pub use current::*;
 pub use entity::{Job, JobEvent, JobType};
 pub use error::JobError;
 pub use es_entity::clock::{Clock, ClockController, ClockHandle};
-pub use handle_api::{JobHandle, JobHandles};
+pub use handle::{JobHandle, JobHandles};
 pub use job_execution::JobStatus;
 pub use migrate::*;
 pub use outcome::{JobOutcome, JobOutcomes, JobReturnValue, JobTerminalState};
